[REFACTOR] adjust folder structure to resolve sonar failure build
