+++ src/components/molecules/navbar/index.test.tsx
@@ -1,7 +1,8 @@
 import { render, screen } from '@testing-library/react'
 
-import Navbar from '@/components/molecules/Navbar.tsx'
 import { withWrappers } from '@/lib/testing/utils.tsx'
+
+import Navbar from './index.tsx'
 
 describe('<Navbar />', () => {
   it('should render kompas gramedia image', () => {
